+++ packages/angular/types/WebtrekkSmartPixelModule.d.ts
@@ -1,0 +1,5 @@
+import { ModuleWithProviders } from '@angular/core';
+import { WebtrekkSmartPixelConfig } from './WebtrekkSmartPixelConfig';
+export declare class WebtrekkSmartPixelModule {
+    static forRoot(conf?: WebtrekkSmartPixelConfig): ModuleWithProviders<WebtrekkSmartPixelModule>;
+}