+++ packages/angular/types/WebtrekkSmartPixelToken.d.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+import { WebtrekkSmartPixelConfig } from './WebtrekkSmartPixelConfig';
+export declare const WEBTREKK_SMART_PIXEL_TOKEN: InjectionToken<WebtrekkSmartPixelConfig>;