+++ packages/angular/types/index.d.ts
@@ -1,0 +1,11 @@
+export { WebtrekkSmartPixelModule } from './WebtrekkSmartPixelModule';
+export { WebtrekkSmartPixelAngular } from './WebtrekkSmartPixelAngular';
+import { WebtrekkSmartPixelModule } from './WebtrekkSmartPixelModule';
+import { WebtrekkSmartPixelAngular } from './WebtrekkSmartPixelAngular';
+export declare const webtrekkSmartPixelAngular: WebtrekkSmartPixelAngular;
+declare const _default: {
+    WebtrekkSmartPixelModule: typeof WebtrekkSmartPixelModule;
+    WebtrekkSmartPixelAngular: typeof WebtrekkSmartPixelAngular;
+    webtrekkSmartPixelAngular: WebtrekkSmartPixelAngular;
+};
+export default _default;