+++ packages/react/shared/src/webtrekkMiddleware.js
@@ -8,12 +8,23 @@ const webtrekkMiddleware = (actions = {}) => ({getState}) => next => action => {
         return nextValue;
     }
 
-    if (!action.webtrekk) {
+    if (!action.webtrekk && action.payload && !action.payload.webtrekk) {
         return nextValue;
     }
 
     var isWebtrekkAction = true;
-    var webtrekk = action.webtrekk;
+
+    var webtrekk;
+    if (action.webtrekk) {
+        webtrekk = action.webtrekk;
+    }
+    else if (action.payload && action.payload.webtrekk) {
+        webtrekk = action.payload.webtrekk;
+    }
+    else {
+        webtrekk = {};
+    }
+
     var type = webtrekk.type || '';
     var data = webtrekk.data || {};
     var sendInstantly = webtrekk.sendInstantly || false;