+++ packages/vue/src/lib/WebtrekkSmartPixelVue.js
@@ -31,6 +31,7 @@ const init_ = function() {
     if (window_ !== null && document_ !== null) {
         pixel_ = wtSmart.use(window_, document_);
         window_['wtSmart'] = pixel_;
+        window_['wtSmart']['_ps'](8, '###VERSION###');
     }
 };