+++ scripts/rollup/build.js
@@ -1,6 +1,8 @@
 const rollup = require('rollup');
 const color = require('./../color');
+const { execSync } = require("child_process");
 
+const root = process.cwd();
 const PACKAGE_NAME = process.argv[2];
 const packages = require('./packages')[PACKAGE_NAME];
 
@@ -28,6 +30,7 @@ async function buildBundle(pkg) {
         for (let i = 0, l = rollupOutputConfig.length; i < l; i++) {
             console.log(color.BLUE, `Start processing "${PACKAGE_NAME}" of ${rollupOutputConfig[i].file}.`);
             await bundle.write(rollupOutputConfig[i]).then(() => {
+                execSync('node ' + root + '/scripts/set-version.js ' + pkg.packagePath, {stdio: 'inherit'});
                 console.log(color.BLUE, `Ends processing "${PACKAGE_NAME}" of ${rollupOutputConfig[i].file}.`);
             });
         }