+++ packages/react/vanilla/CHANGELOG.md
@@ -1,9 +1,17 @@
+<a name="1.2.4"></a>
+## 1.2.4 (2022-09-08)
+
+## Optimization
+- Add tracking parameter (pf, cs801 and cs802) for pixel feature usage
+
+
 <a name="1.2.3"></a>
 ## 1.2.3 (2022-06-09)
 
 ### Fix
 
 * React not working with **react-router-dom** v6
+
 
 <a name="1.2.2"></a>
 ## 1.2.2 (2022-05-04)
@@ -12,12 +20,14 @@
 
 * Limit number of retries to send requests in request queue ([WebtrekkAdvancedData](https://documentation.mapp.com/latest/en/next-15741408.html#id-.SMPNextv1.0-%3CWebtrekkAdvancedData%3E) and [advanced](https://documentation.mapp.com/latest/en/next-15741408.html#id-.SMPNextv1.0-advanced))
 
+
 <a name="1.2.1"></a>
 ## 1.2.1 (2021-12-20)
 
 ### Features
 
 * Add a property to configure request limitation ([WebtrekkAdvancedData](https://documentation.mapp.com/latest/en/react-15741412.html#id-.SMPReactv1.0-%3CWebtrekkAdvancedData%3E) and [advanced](https://documentation.mapp.com/latest/en/react-15741412.html#id-.SMPReactv1.0-advanced))
+
 
 <a name="1.2.0"></a>
 ## 1.2.0 (2021-10-14)
@@ -26,12 +36,14 @@
 
 * Optimize components property types
 
+
 <a name="1.1.3"></a>
 ## 1.1.3 (2021-09-01)
 
 ### Fix
 
 * Change typescript definition (d.ts file).
+
 
 <a name="1.1.2"></a>
 ## 1.1.2 (2020-11-18)
@@ -40,12 +52,14 @@
 
 * Implement Opt-in of tracking cookies 
 
+
 <a name="1.1.0"></a>
 ## 1.1.0 (2020-11-10)
 
 ### Fix
 
 * Restructuring to remove indirect dependency *react-router-dom* for NextJS.
+
 
 <a name="1.0.4"></a>
 ## 1.0.4 (2020-04-14)
@@ -54,20 +68,24 @@
 
 * Support configuration of *sendViaSDK* and *sendViaServer* in \<WebtrekkAdvancedData\>
 
+
 <a name="1.0.3"></a>
 ## 1.0.3 (2019-11-21)
 
 * Renaming package to **@webtrekk-smart-pixel/react**
+
 
 <a name="1.0.2"></a>
 ## 1.0.2 (2019-09-10)
 
 * Compiling build files into ES5 standard
 
+
 <a name="1.0.1"></a>
 ## 1.0.1 (2019-09-09)
 
 * Build separate files for ESM, UMD, AMD and CJS
+
 
 <a name="1.0.0"></a>
 ## 1.0.0 (2019-08-12)
@@ -79,15 +97,18 @@
 * Support react hook reducer ([see](https://documentation.mapp.com/latest/en/react-15741412.html#id-.SMPReactv1.0-webtrekkReducer))
 * Support react redux middleware ([see](https://documentation.mapp.com/latest/en/react-15741412.html#id-.SMPReactv1.0-webtrekkMiddleware))
 
+
 <a name="0.2.5"></a>
 ## 0.2.5 (2019-09-10)
 
 * Compiling build files into ES5 standard
 
+
 <a name="0.2.4"></a>
 ## 0.2.4 (2019-09-09)
 
 * Build separate files for ESM, UMD, AMD and CJS
+
 
 <a name="0.2.3"></a>
 ## 0.2.3 (2019-06-12)
@@ -97,20 +118,24 @@
 * Extension for Content Engagement ([Content Engagement](https://documentation.mapp.com/latest/en/content-engagement-15741314.html))
 * Extension for Cookie Control ([Cookie Control](https://documentation.mapp.com/latest/en/cookie-control-15741322.html))
 
+
 <a name="0.2.2"></a>
 ## 0.2.2 (2019-05-29)
 
 * support 'module.exports'
+
 
 <a name="0.2.1"></a>
 ## 0.2.1 (2019-05-21)
 
 * Add JSDoc
 
+
 <a name="0.2.0"></a>
 ## 0.2.0 (2019-05-20)
 
 * Add example file
+
 
 <a name="0.1.0"></a>
 ## 0.1.0 (2019-05-17)