+++ packages/vue/CHANGELOG.md
@@ -1,9 +1,17 @@
+<a name="1.2.4"></a>
+## 1.2.4 (2022-09-08)
+
+## Optimization
+- Add tracking parameter (pf, cs801 and cs802) for pixel feature usage
+
+
 <a name="1.2.3"></a>
 ## 1.2.3 (2022-06-24)
 
 ### Bugfix
 
 * Fix uncaught ReferenceError: regeneratorRuntime is not defined
+
 
 <a name="1.2.2"></a>
 ## 1.2.2 (2022-05-04)
@@ -12,12 +20,14 @@
 
 * Limit number of retries to send requests in request queue ([advanced](https://documentation.mapp.com/latest/en/vue-15741417.html#id-.SMPVuev1.0-advanced) and [advanced-method](https://documentation.mapp.com/latest/en/vue-15741417.html#id-.SMPVuev1.0-advanced-method))
 
+
 <a name="1.2.1"></a>
 ## 1.2.1 (2021-12-20)
 
 ### Features
 
 * Add a property to configure request limitation ([advanced](https://documentation.mapp.com/latest/en/vue-15741417.html#id-.SMPVuev1.0-advanced) and [advanced-method](https://documentation.mapp.com/latest/en/vue-15741417.html#id-.SMPVuev1.0-advanced-method))
+
 
 <a name="1.2.0"></a>
 ## 1.2.0 (2021-10-14)
@@ -26,12 +36,14 @@
 
 * Optimize property types
 
+
 <a name="1.1.0"></a>
 ## 1.1.3 (2021-09-02)
 
 ### Feature
 
 * Add typescript definition file
+
 
 <a name="1.0.0"></a>
 ## 1.0.0 (2019-11-21)